first step towards control logic flow
Add a flow_control connector to each Element. The purpose is to only
execute this block when flow_control == True. By default (if not
connected) it is True. It can have also the values False and Unknown, in
which case the block should not be executed. This flow_control connector
can also be invalid if the value is still being calculated (which is
different from "unknown", which is a final value).