support for most simple loop structure
To be able to execute some elements more than once, I had to simplify
Flow.run(). Now we don't keep track of the lists elements_to_do and
elements_left. We always iterate over all elements and only keep track
of the number we executed. If we executed 0, we are done.